+++ flink-table/flink-table-common/src/main/java/org/apache/flink/table/factories/FactoryUtil.java
@@ -33,10 +33,6 @@ import org.apache.flink.table.connector.format.DecodingFormat;
 import org.apache.flink.table.connector.format.EncodingFormat;
 import org.apache.flink.table.connector.sink.DynamicTableSink;
 import org.apache.flink.table.connector.source.DynamicTableSource;
-import org.apache.flink.table.types.DataType;
-import org.apache.flink.table.types.logical.LogicalType;
-import org.apache.flink.table.types.logical.LogicalTypeRoot;
-import org.apache.flink.table.types.logical.utils.LogicalTypeChecks;
 import org.apache.flink.table.utils.EncodingUtils;
 import org.apache.flink.util.Preconditions;
 
@@ -54,9 +50,6 @@ import java.util.ServiceConfigurationError;
 import java.util.ServiceLoader;
 import java.util.Set;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-import static org.apache.flink.table.types.logical.utils.LogicalTypeChecks.hasRoot;
 
 /**
  * Utility for working with {@link Factory}s.
@@ -79,6 +72,13 @@ public final class FactoryUtil {
 			"Uniquely identifies the connector of a dynamic table that is used for accessing data in " +
 			"an external system. Its value is used during table source and table sink discovery.");
 
+	public static final ConfigOption<String> FORMAT = ConfigOptions
+		.key("format")
+		.stringType()
+		.noDefaultValue()
+		.withDescription("Defines the format identifier for encoding data. " +
+			"The identifier is used to discover a suitable format factory.");
+
 	public static final ConfigOption<Integer> SINK_PARALLELISM = ConfigOptions
 			.key("sink.parallelism")
 			.intType()
@@ -87,60 +87,14 @@ public final class FactoryUtil {
 					+ "By default, if this option is not defined, the planner will derive the parallelism "
 					+ "for each statement individually by also considering the global configuration.");
 
-	public static final ConfigOption<String> KEY_FORMAT = ConfigOptions
-		.key("key.format")
-		.stringType()
-		.noDefaultValue()
-		.withDescription("Defines the format identifier for encoding key data. " +
-			"The identifier is used to discover a suitable format factory.");
-
-	public static final ConfigOption<String> VALUE_FORMAT = ConfigOptions
-		.key("value.format")
-		.stringType()
-		.noDefaultValue()
-		.withDescription("Defines the format identifier for encoding value data. " +
-			"The identifier is used to discover a suitable format factory.");
-
-	public static final ConfigOption<String> FORMAT = ConfigOptions
-		.key("format")
-		.stringType()
-		.noDefaultValue()
-		.withDescription("Defines the format identifier for encoding data. " +
-			"The identifier is used to discover a suitable format factory.");
-
-	public static final ConfigOption<List<String>> KEY_FIELDS = ConfigOptions
-			.key("key.fields")
-			.stringType()
-			.asList()
-			.defaultValues()
-			.withDescription("Defines an explicit list of physical columns from the table schema "
-					+ "that configure the data type for the key format. By default, this list is "
-					+ "empty and thus a key is undefined.");
-
-	public static final ConfigOption<String> KEY_FIELDS_PREFIX = ConfigOptions
-			.key("key.fields-prefix")
-			.stringType()
-			.noDefaultValue()
-			.withDescription("Defines a custom prefix for all fields of the key format to avoid "
-					+ "name clashes with fields of the value format. By default, the prefix is empty. "
-					+ "If a custom prefix is defined, both the table schema and "
-					+ "'" + KEY_FIELDS.key() + "' will work with prefixed names. When constructing "
-					+ "the data type of the key format, the prefix will be removed and the "
-					+ "non-prefixed names will be used within the key format.");
-
-	public static final ConfigOption<ValueFieldsStrategy> VALUE_FIELDS_INCLUDE = ConfigOptions
-			.key("value.fields-include")
-			.enumType(ValueFieldsStrategy.class)
-			.defaultValue(ValueFieldsStrategy.ALL)
-			.withDescription("Defines a strategy how to deal with key columns in the data type of "
-					+ "the value format. By default, '" + ValueFieldsStrategy.ALL + "' physical "
-					+ "columns of the table schema will be included in the value format which "
-					+ "means that key columns appear in the data type for both the key and value "
-					+ "format.");
-
-	private static final String FORMAT_KEY = "format";
-
-	private static final String FORMAT_SUFFIX = ".format";
+	/**
+	 * Suffix for keys of {@link ConfigOption} in case a connector requires multiple formats (e.g.
+	 * for both key and value).
+	 *
+	 * <p>See {@link #createTableFactoryHelper(DynamicTableFactory, DynamicTableFactory.Context)} for
+	 * more information.
+	 */
+	public static final String FORMAT_SUFFIX = ".format";
 
 	/**
 	 * Creates a {@link DynamicTableSource} from a {@link CatalogTable}.
@@ -229,21 +183,24 @@ public final class FactoryUtil {
 	 *
 	 * <p>The following example sketches the usage:
 	 * <pre>{@code
-	 * // in createDynamicTableSource()
-	 * helper = FactoryUtil.createTableFactoryHelper(this, context);
-	 * keyFormat = helper.discoverScanFormat(DeserializationFormatFactory.class, KEY_FORMAT);
-	 * valueFormat = helper.discoverScanFormat(DeserializationFormatFactory.class, VALUE_FORMAT);
-	 * helper.validate();
-	 * ... // construct connector with discovered formats
+	 *     // in createDynamicTableSource()
+	 *     helper = FactoryUtil.createTableFactoryHelper(this, context);
+	 *
+	 *     keyFormat = helper.discoverEncodingFormat(DeserializationFormatFactory.class, KEY_FORMAT);
+	 *     valueFormat = helper.discoverEncodingFormat(DeserializationFormatFactory.class, VALUE_FORMAT);
+	 *
+	 *     helper.validate();
+	 *
+	 *     ... // construct connector with discovered formats
 	 * }</pre>
 	 *
-	 * <p>Note: The format option parameter of {@code helper.discoverScanFormat(formatFactoryClass, formatOption)}
-	 * and {@code helper.discoverSinkFormat(formatFactoryClass, formatOption)} must be 'format' or
-	 * with '.format' suffix (e.g. {@link #FORMAT}, {@link #KEY_FORMAT} and {@link #VALUE_FORMAT}).
-	 * The discovery logic will replace 'format' with the factory identifier value as the format
-	 * prefix. For example, assuming the identifier is 'json', if format option key is 'format',
-	 * then format prefix is 'json.'. If format option key is 'value.format', then format prefix
-	 * is 'value.json'. The format prefix is used to project the options for the format factory.
+	 * <p>Note: The format option parameter of {@link TableFactoryHelper#discoverEncodingFormat(Class, ConfigOption)}
+	 * and {@link TableFactoryHelper#discoverDecodingFormat(Class, ConfigOption)} must be {@link #FORMAT}
+	 * or end with {@link #FORMAT_SUFFIX}. The discovery logic will replace 'format' with the factory
+	 * identifier value as the format prefix. For example, assuming the identifier is 'json', if the
+	 * format option key is 'format', then the format prefix is 'json.'. If the format option key is
+	 * 'value.format', then the format prefix is 'value.json'. The format prefix is used to project
+	 * the options for the format factory.
 	 *
 	 * <p>Note: This utility checks for left-over options in the final step.
 	 */
@@ -377,92 +334,6 @@ public final class FactoryUtil {
 									.sorted()
 									.collect(Collectors.joining("\n"))));
 		}
-	}
-
-	/**
-	 * Creates an array of indices that determine which physical fields of the table schema to include
-	 * in the key format and the order that those fields have in the key format.
-	 *
-	 * <p>See {@link #KEY_FORMAT} and {@link #KEY_FIELDS} for more information.
-	 */
-	public static int[] createKeyFormatProjection(ReadableConfig options, DataType physicalDataType) {
-		final LogicalType physicalType = physicalDataType.getLogicalType();
-		Preconditions.checkArgument(
-			hasRoot(physicalType, LogicalTypeRoot.ROW),
-			"Row data type expected.");
-		final Optional<String> optionalKeyFormat = options.getOptional(KEY_FORMAT);
-		final Optional<List<String>> optionalKeyFields = options.getOptional(KEY_FIELDS);
-
-		if (!optionalKeyFormat.isPresent() && optionalKeyFields.isPresent()) {
-			throw new ValidationException(
-					String.format(
-							"The option '%s' can only be declared if a key format is defined using '%s'.",
-							KEY_FIELDS.key(),
-							KEY_FORMAT.key()
-					)
-			);
-		} else if (optionalKeyFormat.isPresent() &&
-				(!optionalKeyFields.isPresent() || optionalKeyFields.get().size() == 0)) {
-			throw new ValidationException(
-					String.format(
-							"A key format '%s' requires the declaration of one or more of key fields using '%s'.",
-							KEY_FORMAT.key(),
-							KEY_FIELDS.key()
-					)
-			);
-		}
-
-		if (!optionalKeyFormat.isPresent()) {
-			return new int[0];
-		}
-
-		final List<String> keyFields = optionalKeyFields.get();
-		final List<String> physicalFields = LogicalTypeChecks.getFieldNames(physicalType);
-		return keyFields.stream()
-				.mapToInt(keyField -> {
-					final int pos = physicalFields.indexOf(keyField);
-					if (pos < 0) {
-						throw new ValidationException(
-								String.format(
-										"Could not find the field '%s' in the table schema for usage in the key format. "
-												+ "A key field must be a regular, physical column. "
-												+ "The following columns can be selected in the '%s' option:\n"
-												+ "%s",
-										keyField,
-										KEY_FIELDS.key(),
-										physicalFields
-								)
-						);
-					}
-					return pos;
-				})
-			.toArray();
-	}
-
-	/**
-	 * Creates an array of indices that determine which physical fields of the table schema to include
-	 * in the value format.
-	 *
-	 * <p>See {@link #VALUE_FORMAT} and {@link #VALUE_FIELDS_INCLUDE} for more information.
-	 */
-	public static int[] createValueFormatProjection(ReadableConfig options, DataType physicalDataType) {
-		final LogicalType physicalType = physicalDataType.getLogicalType();
-		Preconditions.checkArgument(
-			hasRoot(physicalType, LogicalTypeRoot.ROW),
-			"Row data type expected.");
-		final int physicalFieldCount = LogicalTypeChecks.getFieldCount(physicalType);
-		final IntStream physicalFields = IntStream.range(0, physicalFieldCount);
-
-		final ValueFieldsStrategy strategy = options.get(VALUE_FIELDS_INCLUDE);
-		if (strategy == ValueFieldsStrategy.ALL) {
-			return physicalFields.toArray();
-		} else if (strategy == ValueFieldsStrategy.EXCEPT_KEY) {
-			final int[] keyProjection = createKeyFormatProjection(options, physicalDataType);
-			return physicalFields
-					.filter(pos -> IntStream.of(keyProjection).noneMatch(k -> k == pos))
-					.toArray();
-		}
-		throw new TableException("Unknown value fields strategy:" + strategy);
 	}
 
 	// --------------------------------------------------------------------------------------------
@@ -705,7 +576,7 @@ public final class FactoryUtil {
 
 		private String formatPrefix(Factory formatFactory, ConfigOption<String> formatOption) {
 			String identifier = formatFactory.factoryIdentifier();
-			if (formatOption.key().equals(FORMAT_KEY)) {
+			if (formatOption.key().equals(FORMAT.key())) {
 				return identifier + ".";
 			} else if (formatOption.key().endsWith(FORMAT_SUFFIX)) {
 				// extract the key prefix, e.g. extract 'key' from 'key.format'
@@ -770,13 +641,6 @@ public final class FactoryUtil {
 		public boolean isTemporary() {
 			return isTemporary;
 		}
-	}
-
-	/**
-	 * Strategies to derive the data type of a value format by considering a key format.
-	 */
-	public enum ValueFieldsStrategy {
-		ALL, EXCEPT_KEY
 	}
 
 	// --------------------------------------------------------------------------------------------
